+++ scripts/mbedtls_framework/psa_test_case.py
@@ -5,10 +5,41 @@
 # SPDX-License-Identifier: Apache-2.0 OR GPL-2.0-or-later
 #
 
-from typing import List, Set
+import os
+import re
+from typing import FrozenSet, List, Optional, Set
 
-from . import psa_information
 from . import test_case
+
+
+# A temporary hack: at the time of writing, not all dependency symbols
+# are implemented yet. Skip test cases for which the dependency symbols are
+# not available. Once all dependency symbols are available, this hack must
+# be removed so that a bug in the dependency symbols properly leads to a test
+# failure.
+def read_implemented_dependencies(filename: str) -> FrozenSet[str]:
+    return frozenset(symbol
+                     for line in open(filename)
+                     for symbol in re.findall(r'\bPSA_WANT_\w+\b', line))
+_implemented_dependencies = None #type: Optional[FrozenSet[str]] #pylint: disable=invalid-name
+
+def find_dependencies_not_implemented(dependencies: List[str]) -> List[str]:
+    """List the dependencies that are not implemented."""
+    global _implemented_dependencies #pylint: disable=global-statement,invalid-name
+    if _implemented_dependencies is None:
+        # Temporary, while Mbed TLS does not just rely on the TF-PSA-Crypto
+        # build system to build its crypto library. When it does, the first
+        # case can just be removed.
+        if os.path.isdir('tf-psa-crypto'):
+            _implemented_dependencies = \
+                read_implemented_dependencies('tf-psa-crypto/include/psa/crypto_config.h')
+        else:
+            _implemented_dependencies = \
+                read_implemented_dependencies('include/psa/crypto_config.h')
+    return [dep
+            for dep in dependencies
+            if (dep.lstrip('!') not in _implemented_dependencies and
+                dep.lstrip('!').startswith('PSA_WANT'))]
 
 
 class TestCase(test_case.TestCase):
@@ -49,6 +80,6 @@ class TestCase(test_case.TestCase):
 
     def skip_if_any_not_implemented(self, dependencies: List[str]) -> None:
         """Skip the test case if any of the given dependencies is not implemented."""
-        not_implemented = psa_information.find_dependencies_not_implemented(dependencies)
+        not_implemented = find_dependencies_not_implemented(dependencies)
         if not_implemented:
             self.add_dependencies(['DEPENDENCY_NOT_IMPLEMENTED_YET'])
